Chrome sometimes throws a DOMException: NotFoundError (8) error.
Don't think this has anything to do with the hashing code. It's more
likely that Chrome doesn't like me using the HTML5 File constructor
to generate fake files.

These changes prevent null from being inserted into [chunks array]
and provide information to the user if this is anything but a testing
environment problem.